Shared Content tab moved to its own page #3522
