+++ src/main/java/net/pms/network/webguiserver/servlets/WebGuiServlet.java
@@ -38,6 +38,7 @@ public class WebGuiServlet extends GuiHttpServlet {
 	public static final String LOGS_BASE_PATH = BASE_PATH + "logs";
 	public static final String PLAYER_BASE_PATH = BASE_PATH + "player";
 	public static final String SETTINGS_BASE_PATH = BASE_PATH + "settings";
+	public static final String SHARED_BASE_PATH = BASE_PATH + "shared";
 
 	private static final List<String> ROUTES = List.of(
 		ABOUT_BASE_PATH,
@@ -45,7 +46,8 @@ public class WebGuiServlet extends GuiHttpServlet {
 		ACTIONS_BASE_PATH,
 		LOGS_BASE_PATH,
 		PLAYER_BASE_PATH,
-		SETTINGS_BASE_PATH
+		SETTINGS_BASE_PATH,
+		SHARED_BASE_PATH
 	);
 
 	@Override
